requirements.txt
numpy==2.3.4
pygame-ce==2.5.5
